requirements.txt
librosa
noisereduce
scipy
numpy
sounddevice
openai-whisper
torch
rapidfuzz
google-genai
python-dotenv
spacy
questionary
